requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
Pillow==10.4.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
WTForms==3.1.2
# Production server
gunicorn==22.0.0
# PostgreSQL (Render / Neon)
psycopg2-binary==2.9.9
# Google OAuth
authlib==1.3.1
requests==2.32.3
